port qmk rgb matrix demo & add openrgb

+++ 2.Firmware/HelloWord-Keyboard-fw/HelloWord/RGBMatrix/rgb_matrix_types.h
@@ -58,7 +58,7 @@ struct RGBCONFIGPACKED {
 } ;
 
 typedef union {
-    uint64_t raw;
+    uint32_t raw;
     struct RGBCONFIGPACKED rgb_config;
 } rgb_config_t;
